KU-104/refactor: PartyCard 컴포넌트 재사용 구조로 분리
